requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
pypdf==5.1.0
